requirements.txt
beautifulsoup4==4.12.3
faiss-cpu==1.8.0.post1
fastapi==0.111.1
httpx==0.27.2
numpy==1.26.4
openai==1.51.2
python-dotenv==1.0.1
uvicorn==0.30.6
